Move postprocessing test case to sima.segment tests
