+++ sima/segment/tests/test_segment.py
@@ -1,7 +1,7 @@
 import sys
 from distutils.version import LooseVersion
 
-import numpy
+import numpy as np
 from numpy.testing import (
     assert_,
     assert_equal,
@@ -14,18 +14,35 @@ from numpy.testing import (
     run_module_suite,
     assert_allclose)
 
-from sima import ImagingDataset
-from sima.misc import example_data
+import os
+import shutil
+
+from sima import ImagingDataset, Sequence, ROI
+from sima.misc import example_data, example_tiff
 from sima import segment
 from sima.segment.ca1pc import cv2_available
 
 
 def setup():
-    return
+    global tmp_dir
+
+    tmp_dir = os.path.join(os.path.dirname(__file__), 'tmp')
+
+    try:
+        os.mkdir(tmp_dir)
+    except:
+        pass
 
 
 def teardown():
-    return
+    global tmp_dir
+
+    shutil.rmtree(tmp_dir)
+
+
+def _gaussian_2d(xy, xy0, xysig):
+    X, Y = np.meshgrid(xy[0], xy[1])
+    return np.exp(-0.5*(((X-xy0[0]) / xysig[0])**2 + ((Y-xy0[1]) / xysig[1])**2))
 
 
 def test_extract_rois():
@@ -34,7 +51,7 @@ def test_extract_rois():
 
 @dec.knownfailureif(
     sys.version_info > (3, 0) and
-    LooseVersion(numpy.__version__) < LooseVersion('1.9.0'))
+    LooseVersion(np.__version__) < LooseVersion('1.9.0'))
 def test_STICA():
     ds = ImagingDataset.load(example_data())
     method = segment.STICA(components=5)
@@ -58,6 +75,74 @@ def test_PlaneCA1PC():
     ds = ImagingDataset.load(example_data())[:, :, :, :50, :50]
     method = segment.PlaneCA1PC(num_pcs=5)
     ds.segment(method)
+
+
+class TestPostprocess(object):
+
+    def setup(self):
+        global tmp_dir
+
+        self.filepath = os.path.join(tmp_dir, "test_imaging_dataset.sima")
+        self.tiff_ds = ImagingDataset(
+            [Sequence.create('TIFF', example_tiff(), 1, 1)],
+            self.filepath)
+
+    def teardown(self):
+        shutil.rmtree(self.filepath)
+
+    def test_postprocess(self):
+        roi_xy = [np.arange(self.tiff_ds.sequences[0].shape[2]),
+                  np.arange(self.tiff_ds.sequences[0].shape[3])]
+
+        rois = ROI.ROIList([
+            ROI.ROI(_gaussian_2d(roi_xy, (10,10), (10,10))),
+            ROI.ROI(_gaussian_2d(roi_xy, (40,40), (10,10))),
+            ROI.ROI(_gaussian_2d(roi_xy, (70,70), (10,10))),
+            ROI.ROI(_gaussian_2d(roi_xy, (100,100), (10,10)))])
+
+        tobool = segment.SparseROIsFromMasks(n_processes=2)
+        smooth = segment.SmoothROIBoundaries(n_processes=2)
+        rois = smooth.apply(tobool.apply(rois))
+
+        refrois = [
+            [np.array([[  1.,   1.,   0.],
+                       [ 28.,   1.,   0.],
+                       [ 30.,   4.,   0.],
+                       [ 30.,  28.,   0.],
+                       [ 26.,  30.,   0.],
+                       [  2.,  30.,   0.],
+                       [  1.,  27.,   0.],
+                       [  1.,   1.,   0.]])],
+            [np.array([[ 30.,  29.,   0.],
+                       [ 57.,  29.,   0.],
+                       [ 60.,  30.,   0.],
+                       [ 60.,  57.,   0.],
+                       [ 59.,  60.,   0.],
+                       [ 31.,  60.,   0.],
+                       [ 29.,  57.,   0.],
+                       [ 29.,  36.,   0.],
+                       [ 30.,  29.,   0.]])],
+            [np.array([[ 60.,  59.,   0.],
+                       [ 87.,  59.,   0.],
+                       [ 90.,  60.,   0.],
+                       [ 90.,  87.,   0.],
+                       [ 89.,  90.,   0.],
+                       [ 61.,  90.,   0.],
+                       [ 59.,  87.,   0.],
+                       [ 59.,  66.,   0.],
+                       [ 60.,  59.,   0.]])],
+            [np.array([[  90.,   89.,    0.],
+                       [ 117.,   89.,    0.],
+                       [ 120.,   90.,    0.],
+                       [ 120.,  117.,    0.],
+                       [ 119.,  120.,    0.],
+                       [  91.,  120.,    0.],
+                       [  89.,  117.,    0.],
+                       [  89.,   96.,    0.],
+                       [  90.,   89.,    0.]])]]
+
+        for refroi, roi in zip(refrois, rois):
+            assert_array_equal(refroi[0], roi.coords[0])
 
 
 if __name__ == "__main__":
